kernel: pmm bitmap uses 32-bit words now 	modified:   kernel/include/memory/vmm.h 	modified:   kernel/memory/pmm.c 	modified:   kernel/memory/vmm.c

+++ kernel/include/memory/vmm.h
@@ -64,7 +64,7 @@ extern page_directory_t* boot_page_dir;
 extern page_directory_t* kernel_page_dir;
 
 // pmm_bitmap defined in pmm.c
-extern uint8_t* pmm_bitmap;
+extern uint32_t* pmm_bitmap;
 extern uint32_t pmm_bitmap_size;
 
 /* function declarations */
